+++ src/search-commons/services/generic-search.adapter.ts
@@ -6,6 +6,7 @@ import {ActionTagForm} from '../../commons/utils/actiontag.utils';
 
 export interface GenericSearchAdapter<R extends Record, F extends GenericSearchForm, S extends GenericSearchResult<R, F>> {
     search(mapper: Mapper, query: any, opts: any): Promise<S>;
+    export(mapper: Mapper, query: any, format: string, opts: any): Promise<string>;
 }
 
 export interface GenericFacetAdapter<R extends Record, F extends GenericSearchForm, S extends GenericSearchResult<R, F>> {